+++ assets/js/session/index.js
@@ -1,21 +1,21 @@
 export default {
   setString(key, value) {
-    window.sessionStorage.setItem(key, value);
+    process.client && window.sessionStorage.setItem(key, value);
   },
   getString(key) {
-    return window.sessionStorage.getItem(key);
+    if (process.client) return window.sessionStorage.getItem(key);
   },
   setObject(key, value) {
-    window.sessionStorage.setItem(key, JSON.stringify(value));
+    process.client && window.sessionStorage.setItem(key, JSON.stringify(value));
   },
   getObject(key) {
-    const str = window.sessionStorage.getItem(key);
+    const str = process.client && window.sessionStorage.getItem(key);
     if (str) {
       return JSON.parse(str);
     }
     return null;
   },
   clear() {
-    window.sessionStorage.clear();
+    process.client && window.sessionStorage.clear();
   }
 };
